webContents: set http useragent settings when overriding user agent
